Refactor send money handlers into utils exports
